requirements.txt
Django==5.2.8
djangorestframework==3.15.2
django-allauth==65.1.0
django-oauth-toolkit==2.3.0
djangorestframework-simplejwt==5.3.1
django-filter==24.3
drf-spectacular==0.27.2
python-dotenv==1.0.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
whitenoise==6.8.2
requests==2.32.3
dj-database-url==2.1.0

